backend/requirements.txt
# Core FastAPI and server dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0

# AI and OpenAI integration
openai==1.3.7
openai-agents==0.1.0

# Database and async support
asyncpg==0.29.0

# HTTP client for Supabase auth
httpx==0.25.2

# Supabase integration
supabase==2.3.0

# PDF generation
reportlab==4.0.7
playwright==1.40.0

# Data validation and serialization
pydantic==2.5.0

# Environment and configuration
python-dotenv==1.0.0

# File upload support
python-multipart==0.0.6

# Streamlit for potential admin interface
streamlit==1.28.1